Fix issues with charm and payload upgrade
Commit 1bdca14 added support for using the UCA
focal-ovn-22.03 pocket by default for new deployments on Focal.

Unfortunately the change missed some required changes for this
to work end to end.

Ensure charm works on Focal without any UCA pocket with the
`openstack-release` package already intalled.  The behavior
of `charms.openstack` is quite different dependning on whether
that information is available or not.

Ensure charm is able to detect availability of and to perform
payload upgrade on change of source.

Ensure charm does not attempt payload upgrade on charm upgrade.
When upgrading from a charm that does not have the `ovn-source`
configuration option, the reactive automatic flag
`config.changed.ovn-source` will fire even though it has not
changed.  Clear the flag in upgrade-charm hook.

Closes-Bug: #1991319
Change-Id: Ia653291632e37a8d9448f05f05d585b36b1b6647