func-tests: COS integration tests
Functional tests were extended to verify integration of
ovn-chassis charm with COS via grafana-agent.

Included cos-lite bundle is expected to be deployed on the
same controller, in a separate model, on top of kubernetes
substrate. It provides cross-model relation offers that
grafana-agent connects to.

Inclusion of the second model and the necessity to run kubernetes
were causing Github runners to run out of disk space, therefore
this change also moves the execution of functional tests to
the self-hosted runners.

func-test-pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/1311
Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>